changed the returned type of Timer in Utils.h

+++ includes/Utils.h
@@ -40,13 +40,13 @@ class Timer
 public:
     Timer() : beg_(clock_::now()) {}
     void Reset() { beg_ = clock_::now(); }
-    double Elapsed() const {
+    unsigned int Elapsed() const {
         return std::chrono::duration_cast<milis_>
             (clock_::now() - beg_).count(); }
 
 private:
     typedef std::chrono::high_resolution_clock clock_;
-    typedef std::chrono::duration<double, std::milli > milis_;
+    typedef std::chrono::duration<unsigned int, std::milli > milis_;
     std::chrono::time_point<clock_> beg_;
 };
